requirements.txt
openai==1.43.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pandas
json
numpy
